requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
greenlet==3.1.1
PyJWT==2.11.0
cryptography==46.0.5
alembic==1.18.4
gunicorn==23.0.0
